Extract split_headers to its own file so it can be unit tested
