+++ packages/adapter-netlify/src/handler.js
@@ -1,5 +1,6 @@
 import './shims';
 import { Server } from '0SERVER';
+import { split_headers } from './headers';
 
 /**
  * @param {import('@sveltejs/kit').SSRManifest} manifest
@@ -54,34 +55,4 @@ function to_request(event) {
 	}
 
 	return new Request(rawUrl, init);
-}
-
-/**
- * Splits headers into two categories: single value and multi value
- * @param {Headers} headers
- * @returns {{
- *   headers: Record<string, string>,
- *   multiValueHeaders: Record<string, string[]>
- * }}
- */
-function split_headers(headers) {
-	/** @type {Record<string, string>} */
-	const h = {};
-
-	/** @type {Record<string, string[]>} */
-	const m = {};
-
-	headers.forEach((value, key) => {
-		if (key === 'set-cookie') {
-			// @ts-expect-error (headers.raw() is non-standard)
-			m[key] = headers.raw()[key];
-		} else {
-			h[key] = value;
-		}
-	});
-
-	return {
-		headers: h,
-		multiValueHeaders: m
-	};
 }
