+++ packages/adapter-netlify/src/headers.js
@@ -1,0 +1,29 @@
+/**
+ * Splits headers into two categories: single value and multi value
+ * @param {Headers} headers
+ * @returns {{
+ *   headers: Record<string, string>,
+ *   multiValueHeaders: Record<string, string[]>
+ * }}
+ */
+export function split_headers(headers) {
+	/** @type {Record<string, string>} */
+	const h = {};
+
+	/** @type {Record<string, string[]>} */
+	const m = {};
+
+	headers.forEach((value, key) => {
+		if (key === 'set-cookie') {
+			// @ts-expect-error (headers.raw() is non-standard)
+			m[key] = headers.raw()[key];
+		} else {
+			h[key] = value;
+		}
+	});
+
+	return {
+		headers: h,
+		multiValueHeaders: m
+	};
+}
